backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.0.0  # Supabase PostgreSQL client (replaces SQLModel)
# sqlmodel>=0.0.14  # Replaced by Supabase
# alembic>=1.13.0  # Migrations handled by Supabase
psycopg2-binary>=2.9.9  # PostgreSQL driver (still needed for some ops)
# asyncpg>=0.29.0  # Not needed with Supabase client

# Authentication & Security
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.6  # Form data parsing

# AI & Analytics
openai>=1.10.0  # GPT-5
pandas>=2.2.0  # Data analysis
numpy>=1.26.0  # Numerical operations

# HTTP Clients for API Integrations
httpx>=0.26.0  # Async HTTP client
requests>=2.31.0  # Sync HTTP client
oauthlib>=3.2.2  # OAuth flows
requests-oauthlib>=1.3.1  # OAuth for requests

# Background Tasks
celery[redis]>=5.3.4  # Task queue
redis>=5.0.1  # Cache & message broker

# Utilities
python-dotenv>=1.0.0  # Environment variables
python-dateutil>=2.8.2  # Date utilities
pytz>=2024.1  # Timezone handling

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0  # Code formatting
ruff>=0.1.0  # Linting
mypy>=1.8.0  # Type checking

# Optional but recommended
email-validator>=2.1.0  # Email validation
pillow>=10.2.0  # Image processing
